Refactor the item and dispenser
-Refactor the item and dispenser
-The player can interact withe item and dispenser
