+++ Assets/DataAssets/DispenserDataObject.asset
@@ -1,0 +1,18 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 202b481ce62e3bf4285a7706de1f1e11, type: 3}
+  m_Name: DispenserDataObject
+  m_EditorClassIdentifier: 
+  _type: 0
+  _itemPrefab: {fileID: 0}
+  _dispenserSprite: {fileID: 0}
+  _itemSprite: {fileID: 0}
