+++ Assets/DataAssets/DispenserDataObject.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 751b76bef96100c439cc0a67b12d43ba
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
