+++ Assets/Scenes/Level_Prototype.unity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 305e0fbf0428bfb4ca9696627ea5abcf
+guid: ed22d289c173fd04bbab8d953d9a8ea7
 DefaultImporter:
   externalObjects: {}
   userData:
