--- Assets/Scripts/Dispenser.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-<<<<<<< HEAD:Assets/Scripts/Dispenser.meta
-guid: 90eccad5304b68e4e9fa0f5ff6624706
-=======
-guid: aa46450c262d0aa449118e52f74b2083
->>>>>>> origin/Refactor_Player:Assets/DataAssets.meta
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
