--- Assets/Scripts/Dispenser/Dispenser.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b259d569b9e4f7f43a00a82ede97cd41
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
