+++ Assets/Scripts/ScriptableObjects/DispenserItem.cs
@@ -4,7 +4,6 @@
 public class DispenserItem : ScriptableObject
 {
     [SerializeField] private DispenserData.Type _type = DispenserData.Type.None;
-    //[SerializeField] private Color color = Color.white; 
     [SerializeField] private GameObject _itemPrefab = null;
     [SerializeField] private Sprite _dispenserSprite = null;
     [SerializeField] private Sprite _itemSprite = null;
@@ -13,10 +12,4 @@ public class DispenserItem : ScriptableObject
     public GameObject ItemPerfab => _itemPrefab;
     public Sprite DispenserSprite => _dispenserSprite;
     public Sprite ItemSprite => _itemSprite;
-
-    //public DispenserData.Type DispenserType { get { return type; } set { type = value; } }
-    //public Color DispenserColor { get { return color; } set { color = value; } }
-    //public GameObject ItemPrefab { get { return itemPrefab; } set { itemPrefab = value; } }
-    //public Sprite sprite { get { return sp; }set { sp = value; } }
-    //public Sprite Objectsprite { get { return Objectsp; } set { Objectsp = value; } }
 }
